alertas en elementos

+++ inventarios/modulos/botiquines/elementos/ajax-grid-data.php
@@ -78,7 +78,7 @@ while ($row = mysqli_fetch_array($query)) {  // preparing an array
 	$nestedData[] = $row["comentarios"];
 	$nestedData[] = '<td><center>
                     <a href="editar.php?id_elementos=' . $row['id_elementos'] . '"  data-toggle="tooltip" title="Editar datos" class="btn btn-sm btn-info"> <i class="fa-solid fa-pen-to-square" style="color: #f2eded;"></i> </a>
-                    <a href="index.php?action=delete&id_elementos=' . $row['id_elementos'] . '"  data-toggle="tooltip" title="Eliminar" class="btn btn-sm btn-danger"> <i class="fa-solid fa-trash-can" style="color: #f2eded;"></i> </a>
+                    <a href="index.php?action=delete&id_elementos=' . $row['id_elementos'] . '" onclick="return confirmacion()"  data-toggle="tooltip" title="Eliminar" class="btn btn-sm btn-danger"> <i class="fa-solid fa-trash-can" style="color: #f2eded;"></i> </a>
                     </center></td>';
 
 	$data[] = $nestedData;
